refactor(contracts): Split into storage/ and events/ modules, add REQ-006 docs
- Move DataKey, PropertyInfo, Distribution types to storage/mod.rs
- Move event emission helpers to events/mod.rs
- Add operational assumptions documentation (REQ-006) as code comments
- Matches suggested file structure from issue C3-006
